pyproject.toml
[tool.poetry]
name = "button-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "button_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.25.1"
python-telegram-bot = "13.14"
log2d = "^0.0.18"


[tool.poetry.group.dev.dependencies]
icecream = "^2.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
